+++ CHANGES.md
@@ -1,5 +1,9 @@
 # Changes
 
+## [codec-0.2.1] - 2020-09-17
+
+* Do not add empty Message section to encoded buffer
+
 ## [codec-0.2.0] - 2020-08-05
 
 * Drop In/OutMessage